refactor: extract edge handlers code
