+++ src/commands/deploy.js
@@ -14,8 +14,8 @@ const isObject = require('lodash.isobject')
 const SitesCreateCommand = require('./sites/create')
 const LinkCommand = require('./link')
 const { NETLIFYDEV, NETLIFYDEVLOG, NETLIFYDEVERR } = require('../utils/logo')
-const uploadEdgeHandlersBundle = require('@netlify/plugin-edge-handlers/src/upload')
-const { statAsync, readFileAsyncCatchError } = require('../lib/fs')
+const { statAsync } = require('../lib/fs')
+const { deployEdgeHandlers } = require('../utils/edge-handlers')
 
 const DEFAULT_DEPLOY_TIMEOUT = 1.2e6
 
@@ -131,82 +131,6 @@ const validateFolders = async ({ deployFolder, functionsFolder, error, log }) =>
   return { deployFolderStat, functionsFolderStat }
 }
 
-const validateEdgeHandlerFolder = async ({ edgeHandlersFolder, error }) => {
-  try {
-    const resolvedFolder = path.resolve(process.cwd(), edgeHandlersFolder || '.netlify/edge-handlers')
-    const stat = await statAsync(resolvedFolder)
-    if (!stat.isDirectory()) {
-      error(`Edge Handlers folder ${edgeHandlersFolder} must be a path to a directory`)
-    }
-    return resolvedFolder
-  } catch (e) {
-    // only error if edgeHandlers was passed as an argument
-    if (edgeHandlersFolder) {
-      if (e.code === 'ENOENT') {
-        return error(`No such directory ${edgeHandlersFolder}!`)
-      }
-      // Improve the message of permission errors
-      if (e.code === 'EACCES') {
-        return error('Permission error when trying to access Edge Handlers folder')
-      }
-      throw e
-    }
-  }
-}
-
-const manifestFilename = 'manifest.json'
-
-const readBundleAndManifest = async ({ edgeHandlersResolvedFolder, error }) => {
-  const manifestPath = path.resolve(edgeHandlersResolvedFolder, manifestFilename)
-  const { content: manifest, error: manifestError } = await readFileAsyncCatchError(manifestPath)
-  if (manifestError) {
-    error(`Could not read Edge Handlers manifest file ${manifestPath}: ${manifestError.message}`)
-  }
-
-  let manifestJson
-  try {
-    manifestJson = JSON.parse(manifest)
-  } catch (e) {
-    error(`Edge Handlers manifest file is not a valid JSON file: ${e.message}`)
-  }
-
-  if (!manifestJson.sha) {
-    error(`Edge Handlers manifest file is missing the 'sha' property`)
-  }
-
-  const bundlePath = path.resolve(edgeHandlersResolvedFolder, manifestJson.sha)
-  const { content: bundle, error: bundleError } = await readFileAsyncCatchError(bundlePath)
-
-  if (bundleError) {
-    error(`Could not read Edge Handlers bundle file ${bundlePath}: ${bundleError.message}`)
-  }
-
-  return { bundle, manifest: manifestJson }
-}
-
-const deployEdgeHandlers = async ({ edgeHandlersFolder, deployId, apiToken, silent, error }) => {
-  const edgeHandlersResolvedFolder = await validateEdgeHandlerFolder({ edgeHandlersFolder, error })
-  if (edgeHandlersResolvedFolder) {
-    try {
-      const spinner = silent
-        ? null
-        : ora({
-            text: `Deploying Edge Handlers from directory ${edgeHandlersResolvedFolder}`,
-          }).start()
-
-      const { bundle, manifest } = await readBundleAndManifest({ edgeHandlersResolvedFolder, error })
-      await uploadEdgeHandlersBundle(bundle, manifest, deployId, apiToken)
-      spinner &&
-        spinner.stopAndPersist({
-          text: `Finished deploying Edge Handlers from directory: ${edgeHandlersResolvedFolder}`,
-          symbol: logSymbols.success,
-        })
-    } catch (e) {
-      error(`Failed deploying Edge Handlers: ${e.message}`)
-    }
-  }
-}
-
 const runDeploy = async ({
   flags,
   deployToProduction,
@@ -245,9 +169,8 @@ const runDeploy = async ({
     }
 
     const draft = !deployToProduction && !alias
-    const branch = alias
     const title = flags.message
-    results = await api.createSiteDeploy({ siteId, title, body: { draft, branch } })
+    results = await api.createSiteDeploy({ siteId, title, body: { draft, branch: alias } })
     const deployId = results.id
 
     const silent = flags.json || flags.silent
